requirements.txt
opencv-python==4.11.0.86
pytz==2025.1
Pillow==11.1.0
requests
pyyaml==6.0.2
onvif-zeep==0.2.12
paho-mqtt==2.1.0
beautifulsoup4
psutil==7.0.0
opencv-python-headless==4.11.0.86
numpy==1.24.3
onnxruntime==1.15.1
imageio-ffmpeg==0.6.0
imagehash==4.3.2
flask==3.1.0
qrcode==8.0